De-inlined complex functions, defaulted constructors and destructors, const correctness
